requirements.txt
flask==3.0.3
pandas==2.2.2
openpyxl==3.1.2
XlsxWriter==3.2.0
requests==2.32.3
python-dotenv==1.0.1
